Editing page content works!
Internally, the system is already implemented as a subobject,
allowing the system to grow to multiple content area's (e.g. sidebar).
The change_view method is largely inspired by the implementation of FeinCMS.

Externally, in the interface this is not yet exposed.
It's only possible to edit the main area of the page.

Time: 14h