Implemented proper layout switching
The individual items are migrated to similar regions,
and input values are stored/restored property.

The new django-wysiwyg functions allow enabling/disabling
the wysiwyg editor on demand.

The height of the tabmain area is set to a fixed size,
until the tabs are reorganized and reloaded. This prevents
accidental flashing of the page footer while the tabs switch.

Time: 2.5h