JavaScript admin refactor part 1: splitting files, adding namespaces
The admin.js has been split up in the various aspects it covers,
to improve code navigation. All files are namespaced now,
making it easier to locate functions.

In event bindings, the direct function reference is passed (so no this.x)
but that's OK for me, considering all functions are called by namespace.

The split also revealed some inter-module-relations,
which can be fixed in a second refactoring step.