Fix absolute imports in ecms_plugins package.
Other packages won't have this problem, it's just that ecms_plugins
is both the preferred name for the top-level package and ecms_plugins.py
file that is part of the module itself.

This is why the import previously failed. Since absolute imports will
become the standard, the names can be kept as is.