Plugin API change: render() method on plugin itself.
From a model-view perspective, it's much nicer to have the render()
method on the plugin class (like Django CMS), and not on the model (like
FeinCMS does).

In the framework it's a bit of a hack, considering that there is a link
back from the model to the plugin now.