Improve the plugin render API; include request object.
Currently the `RequestContext` is not reused, as I wonder why you'd want
to have a context processor in all your plugins. If there is a usecase,
it can be quickly added.

Instead, there is a `EcmsPluginContext` that does include the common
fields like `request`, `STATIC_URL` and `MEDIA_URL` which covers 99% of
the usecases.