Restored Firefox layout switch, store initial value in widget.
The cp_plugins.on_init_layout() provides a hook to detect whether
Firefox has changed the initial layout value, and therefore the system
requires a refresh/fetch of the layout.

The value is now stored in the <select> box, avoiding the need to pass
the value via a JavaScript function.