Fix breadcrumb of change+delete pages.
Instead of overwriting/replacing render_change_form(), a templatetag is
added because the admin functions make it really hard to overwrite 'applabel'.
This makes it much easier to fix the breadcrumb, and avoids code duplication.

The templates are now provided via a 'change_form_template' property.
The other pagetype plugins correctly use the `PageAdmin` instead of
`UrlNodeAdmin`, so they get the proper breadcrumb labels too.